refactor(openomni): split read back executor internals
